GUI_modular/requirements.txt
pyserial>=3.5
pyvisa>=1.11.3
pyvisa-py>=0.5.2